+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.0
+- Added EnkaProfile#system, EnkaProfile#fetchGameAccounts(), EnkaProfile#fetchGameAccount(), and EnkaProfile#fetchBuilds().
+- Added EnkaGameAccount#system, and EnkaGameAccount#fetchBuilds().
+- Better typings with generics
 # 0.2.0
 - Made EnkaSystem non-static.
 - Changed the type of HoyoType to 0 | 1.